pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "neural-cache"
version = "0.1.0"
description = "Production-grade Neural Cache System for Memory-Augmented AI"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24,<2.0",
    "faiss-cpu>=1.7.4",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.2",
    "torch>=2.0.0",
    "sqlite-utils>=3.35",
    "pydantic>=2.0",
    "structlog>=23.0",
    "prometheus-client>=0.17",
    "rich>=13.0",
    "click>=8.1",
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-benchmark>=4.0",
    "openai>=1.0",
    "httpx>=0.24",
    "tenacity>=8.2",
    "orjson>=3.9",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pre-commit",
]
gpu = [
    "faiss-gpu>=1.7.4",
]

[project.scripts]
neural-cache = "neural_cache.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
